Update Customer.cs
Moved Linq call using Lazy object to constructor. Even though orders look like they are being loaded in the constructor, the Lazy object will wait until orders are requested before loading them. It does the checking to see if orders have been loaded so that this does not have to be done manually.